Separate intervals from coverage
